# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Technical Documentation Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DocRecord {
  id: number;
  title: string;
  /** readme | api | functions | setup | changelog | overview */
  docType: string;
  /** file | github */
  sourceType: string;
  /**
     * GitHub URL or filename
     * @nullable
     */
  sourceRef?: string | null;
  /** Generated markdown content */
  content: string;
  /**
     * Detected programming language
     * @nullable
     */
  language?: string | null;
  createdAt: string;
}

export interface DocInput {
  /** @minLength 1 */
  title: string;
  /** readme | api | functions | setup | changelog | overview */
  docType: string;
  /** file | github */
  sourceType: string;
  /** GitHub URL or filename */
  sourceRef?: string;
  /** Raw code content to document */
  codeContent?: string;
  /** Programming language hint */
  language?: string;
}

export type DocStatsByType = {[key: string]: number};

export interface DocStats {
  total: number;
  byType: DocStatsByType;
  thisWeek: number;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}