# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Technical Documentation Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: docs
    description: Documentation generation and history
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /docs:
    get:
      operationId: listDocs
      tags: [docs]
      summary: List all generated documentation records
      responses:
        "200":
          description: Array of doc records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocRecord"
    post:
      operationId: generateDoc
      tags: [docs]
      summary: Generate documentation from uploaded code or GitHub URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocInput"
      responses:
        "201":
          description: Generated documentation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /docs/{id}:
    get:
      operationId: getDoc
      tags: [docs]
      summary: Get a single documentation record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documentation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDoc
      tags: [docs]
      summary: Delete a documentation record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /docs/stats:
    get:
      operationId: getDocStats
      tags: [docs]
      summary: Get dashboard stats (totals by type, recent count)
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocStats"

  /docs/recent:
    get:
      operationId: listRecentDocs
      tags: [docs]
      summary: List the 5 most recently generated docs
      responses:
        "200":
          description: Recent docs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocRecord"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DocRecord:
      type: object
      required: [id, title, docType, sourceType, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        docType:
          type: string
          description: "readme | api | functions | setup | changelog | overview"
        sourceType:
          type: string
          description: "file | github"
        sourceRef:
          type: ["string", "null"]
          description: GitHub URL or filename
        content:
          type: string
          description: Generated markdown content
        language:
          type: ["string", "null"]
          description: Detected programming language
        createdAt:
          type: string
          format: date-time

    DocInput:
      type: object
      required: [title, docType, sourceType]
      properties:
        title:
          type: string
          minLength: 1
        docType:
          type: string
          description: "readme | api | functions | setup | changelog | overview"
        sourceType:
          type: string
          description: "file | github"
        sourceRef:
          type: string
          description: GitHub URL or filename
        codeContent:
          type: string
          description: Raw code content to document
        language:
          type: string
          description: Programming language hint

    DocStats:
      type: object
      required: [total, byType, thisWeek]
      properties:
        total:
          type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
        thisWeek:
          type: integer

    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error