# lib/api-zod/src/generated/types/docInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Technical Documentation Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DocInput {
  /** @minLength 1 */
  title: string;
  /** readme | api | functions | setup | changelog | overview */
  docType: string;
  /** file | github */
  sourceType: string;
  /** GitHub URL or filename */
  sourceRef?: string;
  /** Raw code content to document */
  codeContent?: string;
  /** Programming language hint */
  language?: string;
}