# lib/api-zod/src/generated/types/docRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Technical Documentation Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DocRecord {
  id: number;
  title: string;
  /** readme | api | functions | setup | changelog | overview */
  docType: string;
  /** file | github */
  sourceType: string;
  /**
     * GitHub URL or filename
     * @nullable
     */
  sourceRef?: string | null;
  /** Generated markdown content */
  content: string;
  /**
     * Detected programming language
     * @nullable
     */
  language?: string | null;
  createdAt: Date;
}