# lib/api-zod/src/generated/types/docStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Technical Documentation Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { DocStatsByType } from './docStatsByType';

export interface DocStats {
  total: number;
  byType: DocStatsByType;
  thisWeek: number;
}